Add files via upload
Forms are always an essential part of any project that is used to collect user input information.  we are creating a survey form for  using HTML and CSS. In this form, there are several questions that asked by users. Here, we will use HTML to design the basic layout of the form and CSS will give a beautiful design to our layout like text-decoration, text color, background color, text alignment, margin, padding, etc.
 concepts of HTML and CSS like tags, ID, color, margin, padding, font, text field, checkbox, radio button, etc.
Inside the form element, provide different options to the user like name, email, age, etc. In HTML, we have different input elements for different tasks. So, we are providing respective input types according to the question (email for taking email input, text-area for taking a large message, the name for taking name input, etc.).